update system message
I specified in the system message that messages should be short and concise. While it can be argued that shorter messages aren't as good, they fir on mobile devices much better and the speed is greatly improved. I have seen a reduce in the message response time  by as much as 20 seconds.